server/requirements.txt
pymongo>=3.9
PyYAML>=5.1
dnspython>=1.16
flask-login>=0.4
Werkzeug>=0.16
